test/requirements.txt
autobahn>=0.10.7
autobahntestsuite>=0.7.2
psutil>=3.2.2
pytest>=2.8.1
pytest-twisted>=1.5
unittest2
tap.py>=1.7
yamlish>=0.18.0
